Change format seconds from hh:mm:ss => h:mm:ss
